+++ src/glob.js
@@ -12,26 +12,25 @@ function match(pattern, str) {
  * A helper function to match input against [pattern-list](https://www.freebsd.org/cgi/man.cgi?query=ssh_config&sektion=5#PATTERNS).
  * According to `man ssh_config`, negated patterns shall be matched first.
  *
- * @param {string} patternList
+ * @param {string|string[]} patternList
  * @param {string} str
  */
 function glob(patternList, str) {
-  const patterns = patternList.split(/[,\s]+/).sort((a, b) => {
-    return a.startsWith('!') ? -1 : 1
-  })
+  const patterns = Array.isArray(patternList) ? patternList : patternList.split(/[,\s]+/)
 
+  // > If a negated entry is matched, then the Host entry is ignored, regardless of whether any other patterns on the line match.
+  let result = false
   for (const pattern of patterns) {
     const negate = pattern[0] == '!'
 
     if (negate && match(pattern.slice(1), str)) {
       return false
     } else if (match(pattern, str)) {
-      return true
+      result = true
     }
   }
 
-  return false
+  return result
 }
-
 
 module.exports = glob