+++ test/test.ssh-config.js
@@ -1,209 +1,33 @@
 'use strict'
 
-const expect = require('expect.js')
+const assert = require('assert').strict || require('assert')
 const fs = require('fs')
 const path = require('path')
 const heredoc = require('heredoc').strip
 
 const SSHConfig = require('..')
 
-const { DIRECTIVE, COMMENT } = SSHConfig
+const { DIRECTIVE } = SSHConfig
 
 function readFile(fpath) {
   return fs.readFileSync(path.join(__dirname, fpath), 'utf-8')
     .replace(/\r\n/g, '\n')
 }
 
-
 describe('SSHConfig', function() {
-  it('.parse simple config', function() {
-    let config = SSHConfig.parse(readFile('fixture/config'))
-
-    expect(config[0].param).to.equal('ControlMaster')
-    expect(config[0].value).to.equal('auto')
-    expect(config.length).to.equal(7)
-
-    expect(config.find({ Host: 'tahoe1' })).to.eql({
-      type: DIRECTIVE,
-      before: '',
-      after: '\n',
-      param: 'Host',
-      separator: ' ',
-      value: 'tahoe1',
-      config: [{
-        type: DIRECTIVE,
-        before: '  ',
-        after: '\n',
-        param: 'HostName',
-        separator: ' ',
-        value: 'tahoe1.com',
-      }, {
-        type: DIRECTIVE,
-        before: '  ',
-        after: '\n\n',
-        param: 'Compression',
-        separator: ' ',
-        value: 'yes'
-      }]
-    })
-  })
-
-
-  it('.parse config with parameters and values separated by =', function() {
-    let config2 = SSHConfig.parse(heredoc(function() {/*
-      Host=tahoe4
-        HostName=tahoe4.com
-        User=keanu
-    */}))
-
-    expect(config2[0]).to.eql({
-      type: DIRECTIVE,
-      before: '',
-      after: '\n',
-      param: 'Host',
-      separator: '=',
-      value: 'tahoe4',
-      config: [{
-        type: DIRECTIVE,
-        before: '  ',
-        after: '\n',
-        param: 'HostName',
-        separator: '=',
-        value: 'tahoe4.com'
-      },{
-        type: DIRECTIVE,
-        before: '  ',
-        after: '\n',
-        param: 'User',
-        separator: '=',
-        value: 'keanu'
-      }]
-    })
-  })
-
-
-  it('.parse comments', function() {
-    let config = SSHConfig.parse(heredoc(function() {/*
-      # I'd like to travel to lake tahoe.
-      Host tahoe1
-        HostName tahoe1.com
-
-      # or whatever place it is.
-      # I just need another vocation.
-      Host *
-        IdentityFile ~/.ssh/ids/whosyourdaddy
-    */}))
-
-    expect(config[0].type).to.equal(COMMENT)
-    expect(config[0].content).to.equal("# I'd like to travel to lake tahoe.")
-
-    // The comments goes with sections. So the structure is not the way it seems.
-    expect(config[1].config[1].type).to.equal(COMMENT)
-    expect(config[1].config[1].content).to.equal('# or whatever place it is.')
-  })
-
-
-  it('.parse multiple IdentityFile', function() {
-    let config = SSHConfig.parse(heredoc(function() {/*
-      # Fallback Identify Files
-      IdentityFile ~/.ssh/ids/%h/%r/id_rsa
-      IdentityFile ~/.ssh/ids/%h/id_rsa
-      IdentityFile ~/.ssh/id_rsa
-    */}))
-
-    expect(config[1].param).to.equal('IdentityFile')
-    expect(config[1].value).to.equal('~/.ssh/ids/%h/%r/id_rsa')
-
-    expect(config[2].param).to.equal('IdentityFile')
-    expect(config[2].value).to.equal('~/.ssh/ids/%h/id_rsa')
-
-    expect(config[3].param).to.equal('IdentityFile')
-    expect(config[3].value).to.equal('~/.ssh/id_rsa')
-  })
-
-
-  it('.parse IdentityFile with spaces', function() {
-    let config = SSHConfig.parse(heredoc(function() {/*
-      IdentityFile C:\Users\fname lname\.ssh\id_rsa
-      IdentityFile "C:\Users\fname lname\.ssh\id_rsa"
-    */}))
-
-    expect(config[0].param).to.equal('IdentityFile')
-    expect(config[0].value).to.equal('C:\\Users\\fname lname\\.ssh\\id_rsa')
-
-    expect(config[1].param).to.equal('IdentityFile')
-    expect(config[1].value).to.equal('C:\\Users\\fname lname\\.ssh\\id_rsa')
-  })
-
-
-  it('.parse Host with double quotes', function() {
-    let config = SSHConfig.parse(heredoc(function() {/*
-      Host foo "!*.bar"
-    */}))
-
-    expect(config[0].param).to.equal('Host')
-    expect(config[0].value).to.equal('foo "!*.bar"')
-  })
-
-
-  it('.stringify the parsed object back to string', function() {
-    let fixture = readFile('fixture/config')
-    let config = SSHConfig.parse(fixture)
-    expect(fixture).to.contain(SSHConfig.stringify(config))
-  })
-
-
-  it('.stringify config with white spaces and comments retained', function() {
-    let config = SSHConfig.parse(heredoc(function() {/*
-      # Lake tahoe
-      Host tahoe4
-
-        HostName tahoe4.com
-        # Breeze from the hills
-        User keanu
-    */}))
-
-    expect(SSHConfig.stringify(config)).to.equal(heredoc(function() {/*
-      # Lake tahoe
-      Host tahoe4
-
-        HostName tahoe4.com
-        # Breeze from the hills
-        User keanu
-    */}))
-  })
-
-
-  it('.stringify IdentityFile entries with double quotes', function() {
-    let config = SSHConfig.parse(heredoc(function() {/*
-      Host example
-        HostName example.com
-        User dan
-        IdentityFile "/path to my/.ssh/id_rsa"
-    */}))
-
-    expect(SSHConfig.stringify(config)).to.equal(heredoc(function() {/*
-      Host example
-        HostName example.com
-        User dan
-        IdentityFile "/path to my/.ssh/id_rsa"
-    */}))
-  })
-
-
   it('.compute by Host', function() {
-    let config = SSHConfig.parse(readFile('fixture/config'))
-    let opts = config.compute('tahoe2')
+    const config = SSHConfig.parse(readFile('fixture/config'))
+    const opts = config.compute('tahoe2')
 
-    expect(opts.User).to.equal('nil')
-    expect(opts.IdentityFile).to.eql(['~/.ssh/id_rsa'])
+    assert(opts.User === 'nil')
+    assert.deepEqual(opts.IdentityFile, ['~/.ssh/id_rsa'])
 
     // the first obtained parameter value will be used. So there's no way to
     // override parameter values.
-    expect(opts.ServerAliveInterval).to.eql(80)
+    assert(opts.ServerAliveInterval == 80)
 
     // the computed result is flat on purpose.
-    expect(config.compute('tahoe1')).to.eql({
+    assert.deepEqual(config.compute('tahoe1'), {
       Compression: 'yes',
       ControlMaster: 'auto',
       ControlPath: '~/.ssh/master-%r@%h:%p',
@@ -219,46 +43,58 @@ describe('SSHConfig', function() {
     })
   })
 
-
   it('.compute by Host with globbing', function() {
-    let config2 = SSHConfig.parse(readFile('fixture/config2'))
+    const config = SSHConfig.parse(heredoc(function() {/*
+      Host example*
+        HostName example.com
+        User simon
+    */}))
 
-    expect(config2.compute('example1')).to.eql({
-      Host: 'example1',
-      HostName: 'example1.com',
-      User: 'simon',
-      Port: '1000',
-      IdentityFile: [
-        '/path/to/key'
-      ]
+    assert.deepEqual(config.compute('example1'), {
+      Host: 'example*',
+      HostName: 'example.com',
+      User: 'simon'
     })
   })
 
+  it('.compute by Host with multiple patterns', function() {
+    const config = SSHConfig.parse(heredoc(function() {/*
+      Host foo "*.bar" "baz ham"
+        HostName example.com
+        User robb
+    */}))
+
+    for (const host of ['foo', 'foo.bar', 'baz ham']) {
+      assert.deepEqual(config.compute(host), {
+        Host: ['foo', '*.bar', 'baz ham'],
+        HostName: 'example.com',
+        User: 'robb'
+      })
+    }
+  })
 
   it('.find with nothing shall yield error', function() {
-    let config = SSHConfig.parse(readFile('fixture/config'))
-    expect(function() { config.find() }).to.throwException()
-    expect(function() { config.find({}) }).to.throwException()
+    const config = SSHConfig.parse(readFile('fixture/config'))
+    assert.throws(function() { config.find() })
+    assert.throws(function() { config.find({}) })
   })
-
 
   it('.find shall return null if nothing were found', function() {
-    let config = SSHConfig.parse(readFile('fixture/config'))
-    expect(config.find({ Host: 'not.exist' })).to.be(null)
+    const config = SSHConfig.parse(readFile('fixture/config'))
+    assert(config.find({ Host: 'not.exist' }) == null)
   })
 
-
   it('.find by Host', function() {
-    let config = SSHConfig.parse(readFile('fixture/config'))
+    const config = SSHConfig.parse(readFile('fixture/config'))
 
-    expect(config.find({ Host: 'tahoe1' })).to.eql({
+    assert.deepEqual(config.find({ Host: 'tahoe1' }), {
       type: DIRECTIVE,
       before: '',
       after: '\n',
       param: 'Host',
       separator: ' ',
       value: 'tahoe1',
-      config: [{
+      config: new SSHConfig({
         type: DIRECTIVE,
         before: '  ',
         after: '\n',
@@ -272,45 +108,50 @@ describe('SSHConfig', function() {
         param: 'Compression',
         separator: ' ',
         value: 'yes'
-      }]
+      })
     })
 
-    expect(config.find({ Host: '*' })).to.eql({
+    assert.deepEqual(config.find({ Host: '*' }), {
       type: DIRECTIVE,
       before: '',
       after: '\n',
       param: 'Host',
       separator: ' ',
       value: '*',
-      config: [{
+      config: new SSHConfig({
         type: DIRECTIVE,
         before: '  ',
         after: '\n\n',
         param: 'IdentityFile',
         separator: ' ',
         value: '~/.ssh/id_rsa'
-      }]
+      })
     })
   })
 
-
   it('.remove by Host', function() {
-    let config = SSHConfig.parse(readFile('fixture/config'))
-    let length = config.length
+    const config = SSHConfig.parse(readFile('fixture/config'))
+    const length = config.length
 
     config.remove({ Host: 'no.such.host' })
-    expect(config.length).to.equal(length)
+    assert(config.length === length)
 
     config.remove({ Host: 'tahoe2' })
-    expect(config.find({ Host: 'tahoe2' })).to.be(null)
-    expect(config.length).to.equal(length - 1)
+    assert(config.find({ Host: 'tahoe2' }) == null)
+    assert(config.length === length - 1)
 
-    expect(function() { config.remove() }).to.throwException()
-    expect(function() { config.remove({}) }).to.throwException()
+    assert.throws(function() { config.remove() })
+    assert.throws(function() { config.remove({}) })
   })
 
   it('.append lines', function() {
-    const config = SSHConfig.parse(readFile('fixture/config2'))
+    const config = SSHConfig.parse(heredoc(function() {/*
+      Host example
+        HostName example.com
+        User root
+        Port 22
+        IdentityFile /path/to/key
+    */}))
 
     config.append({
       Host: 'example2.com',
@@ -319,16 +160,16 @@ describe('SSHConfig', function() {
     })
 
     const opts = config.compute('example2.com')
-    expect(opts.User).to.eql('pegg')
-    expect(opts.IdentityFile).to.eql(['~/.ssh/id_rsa'])
-    expect(config.find({ Host: 'example2.com' })).to.eql({
+    assert(opts.User === 'pegg')
+    assert.deepEqual(opts.IdentityFile, ['~/.ssh/id_rsa'])
+    assert.deepEqual(config.find({ Host: 'example2.com' }), {
       type: DIRECTIVE,
       before: '',
       after: '\n',
       param: 'Host',
       separator: ' ',
       value: 'example2.com',
-      config: [{
+      config: new SSHConfig({
         type: DIRECTIVE,
         before: '  ',
         after: '\n',
@@ -342,33 +183,39 @@ describe('SSHConfig', function() {
         param: 'IdentityFile',
         separator: ' ',
         value: '~/.ssh/id_rsa'
-      }]
+      })
     })
   })
 
   it('.append with original identation recognized', function() {
-    const config = SSHConfig.parse(readFile('fixture/config3'))
+    const config = SSHConfig.parse(heredoc(function () {/*
+      Host example1
+        HostName example1.com
+        User simon
+        Port 1000
+        IdentityFile /path/to/key
+    */}).replace(/  /g, '\t'))
 
     config.append({
       Host: 'example3.com',
       User: 'paul'
     })
 
-    expect(config.find({ Host: 'example3.com' })).to.eql({
+    assert.deepEqual(config.find({ Host: 'example3.com' }), {
       type: DIRECTIVE,
       before: '',
       after: '\n',
       param: 'Host',
       separator: ' ',
       value: 'example3.com',
-      config: [{
+      config: new SSHConfig({
         type: DIRECTIVE,
         before: '\t',
         after: '\n',
         param: 'User',
         separator: ' ',
         value: 'paul'
-      }]
+      })
     })
   })
 
@@ -382,7 +229,7 @@ describe('SSHConfig', function() {
       HostName: 'microsoft.com'
     })
 
-    expect('' + config).to.eql(heredoc(function() {/*
+    assert(config.toString() === heredoc(function() {/*
       Host test
         HostName google.com
       Host test2
@@ -398,7 +245,7 @@ describe('SSHConfig', function() {
       HostName: 'example.com'
     })
 
-    expect('' + config).to.eql(heredoc(function() {/*
+    assert(config.toString() === heredoc(function() {/*
       IdentityFile ~/.ssh/id_rsa
       Host test2
         HostName example.com